+++ primos.h
@@ -16,7 +16,7 @@ int gerarPossivelPrimo()
     // Seta a seed para o gerador de números aleatórios de 24 bits
     std::ranlux24 generator(seed);
 
-    unsigned int possivelPrimo = generator();
+    long long possivelPrimo = generator();
 
     if (possivelPrimo < 0)
         possivelPrimo *= -1;
@@ -24,11 +24,11 @@ int gerarPossivelPrimo()
     return possivelPrimo;
 }
 
-unsigned int gerarPossivelPrimoIntervalo(unsigned int vMax)
+long long gerarPossivelPrimoIntervalo(long long vMax)
 {
     // Define um valor mínimo para não correr o risco de gerar valores muito pequenos.
     int minBits = 18;
-    unsigned int m = pow(2, minBits);
+    long long m = pow(2, minBits);
 
     // Pega a seed a partir do relógid do sistema
     unsigned seed = std::chrono::system_clock::now().time_since_epoch().count();
@@ -36,7 +36,7 @@ unsigned int gerarPossivelPrimoIntervalo(unsigned int vMax)
     // Seta a seed para o gerador de números aleatórios de 24 bits
     std::ranlux24 generator(seed);
 
-    unsigned int num;
+    long long num;
     do
     {
         num = generator();
@@ -55,7 +55,7 @@ unsigned int gerarPossivelPrimoIntervalo(unsigned int vMax)
     a -> Números primos
     a ^ p - 1 != 1 (mod p) -> Tem certeza que não é primo
 */
-bool verificaPrimo(int possivelPrimo) // Implementação do teste do pequeno teorema de Fermat
+bool verificaPrimo(long long possivelPrimo) // Implementação do teste do pequeno teorema de Fermat
 {
     int primos[] = {2, 3, 5, 7, 11, 13, 17, 19, 23, 29, 31, 37, 41, 43, 47, 53,
                     59, 61, 67, 71, 73, 79, 83, 89, 97, 101, 103, 107, 109, 113, 127, 131, 137, 139, 149, 151, 157,
@@ -65,9 +65,9 @@ bool verificaPrimo(int possivelPrimo) // Implementação do teste do pequeno teo
                     499, 503, 509, 521, 523, 541, 547, 557, 563, 569, 571, 577, 587, 593, 599, 601, 607, 613, 617,
                     619, 631, 641, 643, 647, 653, 659, 661, 673, 677, 683, 701, 709, 719, 727, 733, 739, 743, 751};
 
-    int size = sizeof(primos) / sizeof(int);
+    long long size = sizeof(primos) / sizeof(int);
 
-    unsigned int a;
+    long long a;
 
     for (int i = 0; i < size; i++)
     {
@@ -80,7 +80,7 @@ bool verificaPrimo(int possivelPrimo) // Implementação do teste do pequeno teo
     return true;
 }
 
-int gerarPseudoPrimo()
+long long gerarPseudoPrimo()
 {
     bool verifica = false;
